requirements.txt
fastapi
uvicorn
websockets
python-dotenv
requests
